Refactor by widening type result_type from long to long long.

+++ include/network-result.h
@@ -6,7 +6,7 @@
 
 namespace Network
 {
-    typedef long result_type;
+    typedef long long result_type;
 
     struct Result
     {
